CRUDs(Employee, OTP Timer)

+++ src/app/Employee/read-employee/read-employee.component.ts
@@ -3,6 +3,7 @@ import { VentrixDBServiceService } from 'src/app/services/ventrix-db-service.ser
 import { BehaviorSubject, Observable } from 'rxjs';
 import { Employee } from 'src/app/shared/Employee';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-read-employee',
@@ -10,8 +11,19 @@ import { Router } from '@angular/router';
   styleUrls: ['./read-employee.component.css']
 })
 export class ReadEmployeeComponent implements OnInit {
+
   employees:any[] = [];
-  constructor(private ventrixdbservice:VentrixDBServiceService, private router: Router) { }
+  p: number = 1;
+  config: any;
+  noOfRows = 10;
+  //Search query
+  query:string = '';
+  constructor(private ventrixdbservice:VentrixDBServiceService, private router: Router) {
+    this.config = {
+      currentPage: 1,
+      itemsPerPage: 2
+    };
+   }
 
   ngOnInit(): void 
   {
@@ -27,6 +39,10 @@ export class ReadEmployeeComponent implements OnInit {
     this.router.navigate(['/create-employee']);
   }
 
+  pageChange(newPage: number) {
+    this.router.navigate(['/read-employee'], {queryParams:{page: newPage}})
+  }
+
   editEmployee(selectedEmployee: Employee)
   {
       this.ventrixdbservice.setEmployee(selectedEmployee);
@@ -36,7 +52,68 @@ export class ReadEmployeeComponent implements OnInit {
   //Delete Client Function 
   deleteEmployee(selectedEmployee: Employee)
   {
-    this.ventrixdbservice.deleteEmployee(selectedEmployee).subscribe();
-    location.reload();
+    Swal.fire({
+      icon: 'warning',
+      title: 'Are you sure you want to delete this employee?',
+      showDenyButton: true,
+      confirmButtonText: 'Yes',
+      denyButtonText: 'No',
+      confirmButtonColor: '#077bff',
+      allowOutsideClick: false,
+      allowEscapeKey: false
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.ventrixdbservice.deleteEmployee(selectedEmployee).subscribe();
+        this.router.navigate(['/read-employee']).then(() => {
+          window.location.reload();
+        });
+      }
+    })
+    
+  }
+
+  searchEmployee()
+  {
+    if(this.query == '' || this.query.replace(/\s/g, '').length==0)
+    {
+      Swal.fire({
+        icon: 'error',
+        title: 'Invalid search',
+        confirmButtonText: 'OK',
+        confirmButtonColor: '#077bff',
+        allowOutsideClick: false,
+        allowEscapeKey: false
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.router.navigate(['/read-employee']).then(()=>{
+            window.location.reload();
+          });
+        }
+      })
+    }
+    else
+    {
+      this.ventrixdbservice.searchEmployee(this.query.toString()).subscribe(response => {
+        this.employees = response;
+        if(this.employees.length ==0)
+        {
+          Swal.fire({
+            icon: 'error',
+            title: 'No Results Found',
+            confirmButtonText: 'OK',
+            confirmButtonColor: '#077bff',
+            allowOutsideClick: false,
+            allowEscapeKey: false
+          }).then((result) => {
+            if(result.isConfirmed) {
+              this.router.navigate(['/read-employee']).then(() => {
+                window.location.reload();
+              });
+            }
+          })
+        }
+        console.log(this.employees)
+      })
+    }
   }
 }
